Update qemu parameters for newer releases
This commit changes the qemu parameters used to lauch the Virtual
Machine in order to provide a more stable experience with the current
updates from Ubuntu 22.04 and 20.04.

CDROM images now are mounted as read-only, and all disks now use the
virtualised PCI/PCIe bus.

The Virtual Machine now uses a virtualised Q35-based motherboard with a
single core Intel Haswell processor. This avoids surprises that may
happen on future QEMU updates, as the default virtualised platform for
x86_64 VMs may change on future releases.

The virtualised GPU now is able to use host based acceleration using
Mesa. This feature may be enabled on the display configuration with the
parameter `gl=on`.