Fix incompatibility between Packages.gz and VM script
The generated Packages.gz file expected the files to be placed in the
`packages` directory. This conflicted with the virtual machine external
repository configuration, which places the files in the repository's
root directory.

The aforementioned problem has been dealt with, and such action required
the substitution of the relative directory approach to a directory
changing one.