Rewrite scripts unifying their configuration
The scripts relied on environment variables spread across them that had
to be carefully modified to avoid breaking them.

These "gotchas" have been eliminated due to centralisation of all the
variables and functions in a single script that is sourced by the
others: `env-vars.sh`, shipped in this repository as `env-vars.sh.ex`.

The README.md has been updated to document the necessary modifications
to the `env-vars.sh.ex` file.